eos-download-image: attempt to handle truncated responses
Mazen's [seeing][1] partial downloads of images, followed by (of course)
GPG verification failing. I'm not sure why; my hypothesis is that
CloudFront is serving a short response because its connection to S3
drops. This is supported by [Amazon's documentation][0]:

> **Dropped TCP Connections**
>
> If the TCP connection between CloudFront and your origin drops while
> your origin is returning an object to CloudFront, CloudFront behavior
> depends on whether your origin included a Content-Length header in the
> response:
>
> - Content-Length header – CloudFront returns the object to the viewer
>   as it gets the object from your origin. However, if the value of the
>   Content-Length header doesn't match the size of the object, CloudFront
>   doesn't cache the object.

So add a check that each request returned the full amount of data
expected; if it does not, push a new range back onto the queue.

[0]: https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/RequestAndResponseBehaviorCustomOrigin.html
[1]: https://phabricator.endlessm.com/P547